Devices keys filenames are no longer meaningful.
Key files used to be stored in a directory named after the device slug
in a file also named after the same device slug. The total length is
more than 160 characters depending on the organization name.

Key files are now stored directly in the devices directory using the
device slughash and the `.keys` extension. The total length is 37
characters.

# Conflicts:
#	parsec/core/local_device.py